autoBotAPP/requirements.txt
Flask
cloudinary
python-telegram-bot
python-telegram-bot==20.0
